setFilterByAuthorizedAccounts(true) causes hard-to-debug exception
When using `setFilterByAuthorizedAccounts(true)` and the testing device does not have a Google account that had previously signed-in to the app, `GoogleSignIn_Show()` will fail with the exception `GoogleSignIn onFailure: 16: Cannot find a matching credential`, because there are no authorized Google Accounts available. The current implementation does not have handling of this exception as Google's documentation specified, where the app can use `setFilterByAuthorizedAccounts(true)` first, but if there are no authorized Google Accounts available, the app should then use `setFilterByAuthorizedAccounts(false)`. See https://developer.android.com/identity/sign-in/credential-manager-siwg#instantiate-google

I propose that we can just use `setFilterByAuthorizedAccounts(false)`. The upside is that we can avoid the exception, which is very confusing to debug as many developers have experienced:
https://forum.gamemaker.io/index.php?threads/google-sign-in-extension-log-in-trouble.101964/#post-687453
https://stackoverflow.com/questions/61768804/getting-16-cannot-find-a-matching-credential-when-doing-one-tap-sign-in-and-s/61768805#61768805

The downside is that the Google Sign In UI will show account selector every time instead of automatically sign-in. This may be annoying to the users.

I think this change can massively improve the developer experience at a slight expense of the user experience. As the GMC thread shows, because the exception was so hard to debug, the thread owner blue10 decided to abandon the implementation all together. 

Alternatively, we can follow Google's instruction to try `setFilterByAuthorizedAccounts(true)` first and then `setFilterByAuthorizedAccounts(false)`. We will have to cache the exception result from `setFilterByAuthorizedAccounts(true)` and to decide whether to use `setFilterByAuthorizedAccounts(false)` next.